backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
python-multipart==0.0.12
huggingface-hub>=0.20.0
transformers>=4.30.0
nltk>=3.8.0
torch>=2.0.0
